+++ dll/ime/msctfime/msctfime.h
@@ -24,12 +24,13 @@
 #include <shlwapi.h>
 #include <strsafe.h>
 
-#include <cicero/cicbase.h>
-#include <cicero/cicarray.h>
-#include <cicero/cicimc.h>
-#include <cicero/cictf.h>
-#include <cicero/ciccaret.h>
-#include <cicero/cicuif.h>
+#include <cicbase.h>
+#include <cicarray.h>
+#include <cicimc.h>
+#include <cictf.h>
+#include <ciccaret.h>
+#include <cicuif.h>
+#include <cicutb.h>
 
 #include <wine/debug.h>
