+++ dll/win32/msctf/CMakeLists.txt
@@ -29,8 +29,8 @@ add_library(msctf MODULE
     version.rc
     ${CMAKE_CURRENT_BINARY_DIR}/msctf.def)
 
-set_module_type(msctf win32dll)
-target_link_libraries(msctf uuid wine)
+set_module_type(msctf win32dll UNICODE)
+target_link_libraries(msctf uuid wine cicero)
 add_importlibs(msctf user32 advapi32 advapi32_vista msvcrt kernel32 ntdll)
 add_delay_importlibs(msctf ole32 oleaut32)
 add_pch(msctf precomp.h SOURCE)
