+++ dll/win32/msutb/precomp.h
@@ -9,7 +9,6 @@
 
 #include <stdlib.h>
 
-#define WIN32_NO_STATUS
 #define COBJMACROS
 #define INITGUID
 
@@ -25,11 +24,10 @@
 #include <atlbase.h>
 #include <atlcom.h>
 #include <strsafe.h>
-#undef STATUS_NO_MEMORY
 
 #include "resource.h"
-#include <cicero/cicreg.h>
-#include <cicero/cicutb.h>
-#include <cicero/cicuif.h>
+#include <cicreg.h>
+#include <cicutb.h>
+#include <cicuif.h>
 
 #include <wine/debug.h>
