+++ sdk/lib/CMakeLists.txt
@@ -6,6 +6,7 @@ add_subdirectory(inflib)
 if(CMAKE_CROSSCOMPILING)
 
 add_subdirectory(3rdparty)
+add_subdirectory(cicero)
 add_subdirectory(comsupp)
 add_subdirectory(conutils)
 add_subdirectory(cportlib)
