+++ sdk/lib/cicero/precomp.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <windows.h>
+#include <unknwn.h>
+#include <stddef.h>
+#include <stdlib.h>
